feat: Extend the DAO pattern to work with multiple databases.
- Add DAOs and DOA factories to support MySQL and Oracle databases.
- Update controller to enable database selection for application interaction.
- Updates gitignore file
